+++ postgres/database-load/locals.tf
@@ -1,0 +1,18 @@
+locals {
+  tags = {
+    application         = var.application
+    environment         = var.environment
+    managed-by          = "DBT Platform - Terraform"
+    copilot-application = var.application
+    copilot-environment = var.environment
+  }
+
+  task_name = "${var.application}-${var.environment}-${var.database_name}-load"
+
+  dump_task_name     = "${var.application}-${var.task.from}-${var.database_name}-dump"
+  dump_kms_key_alias = "alias/${local.dump_task_name}"
+  dump_bucket_name   = local.dump_task_name
+
+  ecr_repository_arn = "arn:aws:ecr-public::763451185160:repository/database-copy"
+
+}